pyproject.toml
[project]
name = "dyno-wallpaper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.3.0",
    "watchdog>=6.0.0",
]
